Reduce array allocations
Object#instance_values is used extensively in the axlsx codebase.
This method is a monkey patch and allocates an array per instance
variable which are immediately passed to Hash::[] and thrown
away. By hand-rolling the behavior of Object#instance_values in
particularly hot code paths we can significatly reduce array
allocation and call to Object#instance_variable_get which aren't
necessary.